requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cssselect==1.2.0
cssutils==2.11.1
Django==5.1.4
django-admin-interface==0.29.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-colorfield==0.11.0
django-email-tools==0.1.1
django-filter==24.3
django-grappelli==4.0.1
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-js-asset==3.0.1
django-material==1.12.0
django-modeltranslation==0.19.12
django-rosetta==0.10.1
django-semantic-admin==0.6.5
django-semantic-forms==0.1.8
django-suit==0.2.25
django-unfold==0.46.0
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
Markdown==3.7
more-itertools==10.6.0
pillow==11.0.0
polib==1.2.0
premailer==3.10.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
